# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Web Umkm",
        "description": "This For API System,And You Can Try It",
        "contact": {
            "email": "muhammadalghifari4321@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/category": {
            "get": {
                "tags": [
                    "Projects UMKM"
                ],
                "summary": "Looking Data Category",
                "description": "Looking Data Category",
                "operationId": "getCategory",
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Projects UMKM"
                ],
                "summary": "Insert Data Category Admin",
                "description": "Insert Data Category Admin",
                "operationId": "insertCategory",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "category": {}
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/category/{id}": {
            "get": {
                "tags": [
                    "Projects UMKM"
                ],
                "summary": "Get By Id Data Category",
                "description": "Get By Id Data Category",
                "operationId": "getByIdCategory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Enter Id Data For Get Data And Relation",
                        "required": true
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Projects UMKM"
                ],
                "summary": "Update Data Category Admin",
                "description": "Update Data Category Admin",
                "operationId": "updateCategory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Enter Id Data For Update",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "category": {}
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Projects UMKM"
                ],
                "summary": "Delete Data Category Admin",
                "description": "Delete Data Category Admin",
                "operationId": "deleteCategory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Enter Id Data For Delete",
                        "required": true
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/compare/barang": {
            "get": {
                "tags": [
                    "Projects UMKM"
                ],
                "summary": "Compare Feature Barang Admin",
                "description": "Compare Feature Barang Admin",
                "operationId": "compareFeature",
                "parameters": [
                    {
                        "name": "barang_id",
                        "in": "query",
                        "description": "Id Barang",
                        "required": true
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/compare": {
            "post": {
                "tags": [
                    "Projects UMKM"
                ],
                "summary": "Compare Price Barang Admin",
                "description": "Compare Price Barang Admin",
                "operationId": "comparePrice",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "barang_id": {},
                                    "first_toko_id": {},
                                    "second_toko_id": {}
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use a global client_id / client_secret and your username / password combo to obtain a token",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "bearer": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    ]
}